A version with a library and two binaries and all three types of tests.
Changes to the description language in files .context and .flow.
Structures defined that are aligned with those definitions.
Hacks to get a first version to compile to learn about modules and tests.
Defined a module structure in the library.
Added one unit test.
Added one integration test.
Added one documentation test in the library.